// src/Store/Ducks/Products/types.ts
export enum ProductsActionTypes {
  GET_PRODUCTS_REQUEST = 'GET_PRODUCTS_REQUEST',
  GET_PRODUCTS_SUCCESS = 'GET_PRODUCTS_SUCCESS',
  GET_PRODUCTS_FAILURE = 'GET_PRODUCTS_FAILURE',
  DELETE_PRODUCT_REQUEST = 'DELETE_PRODUCT_REQUEST',
  DELETE_PRODUCT_SUCCESS = 'DELETE_PRODUCT_SUCCESS',
  DELETE_PRODUCT_FAILURE = 'DELETE_PRODUCT_FAILURE',
  POST_PRODUCT_REQUEST = 'POST_PRODUCT_REQUEST',
  POST_PRODUCT_SUCCESS = 'POST_PRODUCT_SUCCESS',
  POST_PRODUCT_FAILURE = 'POST_PRODUCT_FAILURE',
}